Fix #31: remove sites directory.
It's not clear quite why these are here, and they certainly aren't testcases
per-se (what's their pass condition?!). I think originally to have some sort
of shared files for testing performance on? It also seems unlikely we have
the copyright holders' permission to distribute them under the MIT license in
the case of the Google results and the Python docs.