Fix EOF error line/columns, add to #errors
Fix the line and column numbers.

Currently, the number of errors in the `#errors` section of the test is
the correct number of errors. Counting the ones in `#new-errors` breaks
hundreds of tests because they contain an equivalent error in `#errors`.
So this adds the eof in comment error to `#errors` as well.