ci: deploy docs to github pages
This patch adds a new job to the GitHub workflow to generate docs and deploy them to GH
pages.

The job is only triggered when the PR is merged into the "main" branch. It needs access
to the "GITHUB_TOKEN" of the repo owner. It will push the "target/doc" directory
to a branch called "gh-pages" in the roo directory.

The repo owner will need to update the repo settings to make GitHub deploy to
this directory.