Added documentBuilder struct
Currently, creating a document must be done in 2 steps:

1. create the document
2. create the root node in the document

This is a 2-step process, as we cannot create a root node inside the
constructor of the document.

We moved this to a factory/builder pattern where we can now simply call:

DocumentBuilder::new_document()

and which will correctly create the document AND set the root node.

Also, we have moved the document to the new() constructor instead of the
parse() function. Basically, we should not even need to initialize the
parser, as it's a one-time function (we cannot call the parser twice,
since it will be in a non-init state).