Changing the bytes/CharIterator to the ByteStream
This commit mostly removes all the bytes::CharIterator code and replaces
it with the ByteStream. This is a more clean setup of reading a stream
of characters taken from a stream of bytes. Now both the HTML5 and CSS3
parser will work with the same bytestream system which makes things a
little more clear in the end.