Precedence and groups seems to be working correct
Issues solved with juxtaposition which didn't match correctly.
Also, fixed precedence which was in the wrong order. THe only
thing still needs to do is fix the tests so they not only compile
the syntax, but also check the results (it didn't, so we didn't
catch the errors). It's a bit hard to do right now at my vacation
address with slow wifi, i'm longing for the days of a 2400 baud modem.