Removed nom-tracer as it is not used and relied on insecure package
It relied on lexical-core 0.7 which has some insecurities. Since this
package is used by nom 5 and nom-tracer is the only package that uses
nom-5, we removed the tracer package altogehter (we don't use it
anymore).

This will remove all the references to nom-5 and thus the lexical-core
package.

See: https://github.com/gosub-io/gosub-engine/security/dependabot/5